Refactor: split main.go
